feat: Ajustar velocidad de movimiento en función processGesture
This commit adjusts the movement speed in the `processGesture` function based on the gesture name. If the gesture name is the same as the last gesture, the movement speed is increased. Otherwise, the movement speed is reset to its initial value. This change allows for more precise and controlled mouse movements when interacting with the system.